+++ package.json
@@ -1,0 +1,46 @@
+{
+    "name": "mk-v2",
+    "version": "0.1.0",
+    "private": true,
+    "scripts": {
+      "dev": "next dev",
+      "build": "next build",
+      "start": "next start",
+      "lint": "next lint"
+    },
+    "dependencies": {
+      "@radix-ui/react-avatar": "^1.1.1",
+      "@radix-ui/react-dialog": "^1.1.2",
+      "@radix-ui/react-icons": "^1.3.1",
+      "@radix-ui/react-label": "^2.1.0",
+      "@radix-ui/react-popover": "^1.1.2",
+      "@radix-ui/react-progress": "^1.1.0",
+      "@radix-ui/react-scroll-area": "^1.2.0",
+      "@radix-ui/react-slider": "^1.2.1",
+      "@radix-ui/react-slot": "^1.1.0",
+      "@radix-ui/react-tabs": "^1.1.1",
+      "class-variance-authority": "^0.7.0",
+      "clsx": "^2.1.1",
+      "date-fns": "^3.6.0",
+      "lucide-react": "^0.454.0",
+      "next": "^15.0.2",
+      "next-themes": "^0.3.0",
+      "react": "^18.3.1",
+      "react-day-picker": "8.10.1",
+      "react-dom": "^18.3.1",
+      "shadcn": "^1.0.0",
+      "shadcn-ui": "^0.2.3",
+      "tailwind-merge": "^2.5.4",
+      "tailwindcss-animate": "^1.0.7"
+    },
+    "devDependencies": {
+      "@types/node": "^20",
+      "@types/react": "^18",
+      "@types/react-dom": "^18",
+      "eslint": "^8",
+      "eslint-config-next": "15.0.2",
+      "postcss": "^8",
+      "tailwindcss": "^3.4.1",
+      "typescript": "^5"
+    }
+}
